Move Localization (CLDR) from PrestaShopBundle to Core
