+++ src/Core/Localization/Number/Formatter.php
@@ -24,14 +24,14 @@
  * International Registered Trademark & Property of PrestaShop SA
  */
 
-namespace PrestaShopBundle\Localization\Number;
+namespace PrestaShop\PrestaShop\Localization\Number;
 
 use InvalidArgumentException as SPLInvalidArgumentException;
 use PrestaShop\Decimal\Number as DecimalNumber;
 use PrestaShop\Decimal\Operation\Rounding;
-use PrestaShopBundle\Localization\Exception\LocalizationException;
-use PrestaShopBundle\Localization\Specification\NumberInterface as NumberSpecification;
-use PrestaShopBundle\Localization\Specification\Price as PriceSpecification;
+use PrestaShop\PrestaShop\Localization\Exception\LocalizationException;
+use PrestaShop\PrestaShop\Localization\Specification\NumberInterface as NumberSpecification;
+use PrestaShop\PrestaShop\Localization\Specification\Price as PriceSpecification;
 
 /**
  * Formats a number (raw, price, percentage) according to passed specifications
